refactor: remove RestrictedPython — replace with plain exec in gVisor sandbox
The gVisor container + L3/L4 nftables egress + L7 Envoy forward proxy is
the real tenant boundary. RestrictedPython was not a real boundary (it is
bypassable), redundant inside the sandbox, and caused unnecessary friction
(allow-list rejecting valid imports, RP-specific syntax restrictions).

Changes:
- Delete src/orcheo/graph/ingestion/sandbox.py; move validate_script_size,
  execution_timeout, and _compile_langgraph_script (now plain compile +
  lru_cache) into loader.py
- _execute_langgraph_script uses plain exec with real builtins and
  __name__="__orcheo_ingest__" (preserves if __name__=="__main__" guard)
- _ensure_plugin_sys_path moved to loader.py so orcheo_plugin_* imports
  still resolve without the RP hook
- builder.py passes max_script_bytes=None so size cap is not re-checked on
  every execution rebuild (ingest-only guardrail)
- __init__.py stripped of RP indirection (compile_restricted alias,
  _resolve_compiler, sys/threading/time monkeypatch targets)
- Remove RestrictedPython>=7.2 from pyproject.toml
- Add nft-init one-shot compose service that applies sandbox-egress.nft
  before sandbox-runtime starts, automating the previously manual host step
- Fix misleading deploy doc comment (sandboxing is on by default)
