+++ src/orcheo/graph/ingestion/loader.py
@@ -1,11 +1,21 @@
 """Load LangGraph StateGraph instances from Python scripts."""
 
 from __future__ import annotations
+import ast
 import asyncio
+import builtins
+import contextlib
 import inspect
-from collections.abc import Awaitable
+import signal
+import sys
+import threading
+import time
+from collections.abc import Awaitable, Callable, Generator
 from concurrent.futures import ThreadPoolExecutor
-from typing import Any, get_origin
+from functools import lru_cache
+from pathlib import Path
+from types import CodeType, FrameType
+from typing import Any, cast, get_origin
 from langgraph.graph import StateGraph
 from langgraph.graph.state import CompiledStateGraph
 from orcheo.graph.ingestion.config import (
@@ -13,25 +23,112 @@ from orcheo.graph.ingestion.config import (
     DEFAULT_SCRIPT_SIZE_LIMIT,
 )
 from orcheo.graph.ingestion.exceptions import ScriptIngestionError
-from orcheo.graph.ingestion.sandbox import (
-    compile_langgraph_script,
-    create_sandbox_namespace,
-    execution_timeout,
-    validate_script_size,
-)
 
 
-def _format_syntax_error_message(exc: SyntaxError) -> str:
-    """Format a SyntaxError into a user-friendly message."""
-    if exc.args and isinstance(exc.args[0], str):
-        # Format: "Line 39: AnnAssign statements are not allowed."
-        # or tuple with multiple error messages
-        if isinstance(exc.args[0], str) and exc.args[0].startswith("Line"):
-            return f"Compilation error: {exc.args[0]}"
-        # Handle tuple of errors
-        error_messages = ", ".join(str(arg) for arg in exc.args if isinstance(arg, str))
-        return f"Compilation error: {error_messages}"
-    return f"Compilation error: {exc}"
+TraceFunc = Callable[[FrameType | None, str, object], object]
+
+
+def validate_script_size(source: str, max_script_bytes: int | None) -> None:
+    """Raise ``ScriptIngestionError`` when the script exceeds the byte limit."""
+    if max_script_bytes is None:
+        return
+
+    if max_script_bytes <= 0:
+        msg = "LangGraph script size limit must be a positive integer"
+        raise ScriptIngestionError(msg)
+
+    encoded_length = len(source.encode("utf-8"))
+    if encoded_length > max_script_bytes:
+        msg = f"LangGraph script exceeds the permitted size of {max_script_bytes} bytes"
+        raise ScriptIngestionError(msg)
+
+
+@contextlib.contextmanager
+def execution_timeout(
+    timeout_seconds: float | None,
+    *,
+    sys_module: Any | None = None,
+    threading_module: Any | None = None,
+    time_module: Any | None = None,
+) -> Generator[None, None, None]:
+    """Enforce a wall-clock timeout around script execution."""
+    if timeout_seconds is None or timeout_seconds <= 0:
+        yield
+        return
+
+    sys_obj = sys_module or sys
+    threading_obj = threading_module or threading
+    time_obj = time_module or time
+
+    use_signal = (
+        hasattr(signal, "setitimer")
+        and threading_obj.current_thread() is threading_obj.main_thread()
+    )
+
+    if use_signal:
+        previous_handler = signal.getsignal(signal.SIGALRM)
+
+        def _handle_timeout(_signum: int, _frame: FrameType | None) -> None:
+            raise TimeoutError(
+                "LangGraph script execution timed out"
+            )  # pragma: no cover
+
+        try:
+            signal.signal(signal.SIGALRM, _handle_timeout)
+            signal.setitimer(signal.ITIMER_REAL, timeout_seconds)
+            yield
+        finally:
+            signal.setitimer(signal.ITIMER_REAL, 0)
+            signal.signal(signal.SIGALRM, previous_handler)
+        return
+
+    deadline = time_obj.perf_counter() + timeout_seconds
+
+    def _trace_timeout(_frame: FrameType | None, event: str, _arg: object) -> TraceFunc:
+        if event == "line" and time_obj.perf_counter() > deadline:
+            raise TimeoutError("LangGraph script execution timed out")
+        return _trace_timeout
+
+    previous_trace = cast(TraceFunc | None, sys_obj.gettrace())
+    previous_thread_trace = cast(TraceFunc | None, threading_obj.gettrace())
+
+    sys_obj.settrace(cast(Any, _trace_timeout))
+    threading_obj.settrace(cast(Any, _trace_timeout))
+    try:
+        yield
+    finally:
+        if previous_trace is None:
+            sys_obj.settrace(cast(Any, None))
+        else:
+            sys_obj.settrace(cast(Any, previous_trace))
+
+        if previous_thread_trace is None:
+            threading_obj.settrace(cast(Any, None))
+        else:
+            threading_obj.settrace(cast(Any, previous_thread_trace))
+
+
+def _ensure_plugin_sys_path() -> None:
+    """Expose the managed plugin site-packages directory to imports."""
+    from orcheo.plugins.paths import build_storage_paths
+
+    install_dir = Path(build_storage_paths().install_dir)
+    site_packages = (
+        install_dir
+        / "lib"
+        / f"python{sys.version_info.major}.{sys.version_info.minor}"
+        / "site-packages"
+    )
+    if site_packages.exists() and str(site_packages) not in sys.path:
+        sys.path.insert(0, str(site_packages))
+
+
+@lru_cache(maxsize=128)
+def _compile_langgraph_script(source: str) -> CodeType:
+    """Compile a LangGraph script with top-level await support, with caching."""
+    return compile(
+        source, "<langgraph-script>", "exec", flags=ast.PyCF_ALLOW_TOP_LEVEL_AWAIT
+    )
 
 
 def _execute_langgraph_script(
@@ -39,22 +136,23 @@ def _execute_langgraph_script(
     max_script_bytes: int | None,
     execution_timeout_seconds: float | None,
 ) -> dict[str, Any]:
-    """Execute the LangGraph script inside the sandbox and return its namespace."""
+    """Execute the LangGraph script and return its namespace."""
     validate_script_size(source, max_script_bytes)
-    namespace = create_sandbox_namespace()
-
+    _ensure_plugin_sys_path()
+    namespace: dict[str, Any] = {
+        "__name__": "__orcheo_ingest__",
+        "__builtins__": builtins,
+    }
     try:
-        compiled = compile_langgraph_script(source)
+        compiled = _compile_langgraph_script(source)
         with execution_timeout(execution_timeout_seconds):
-            exec(compiled, namespace)
+            exec(compiled, namespace)  # noqa: S102
     except ScriptIngestionError:
         raise
     except SyntaxError as exc:
-        # RestrictedPython syntax errors come with detailed information
-        message = _format_syntax_error_message(exc)
+        message = f"Compilation error: {exc}"
         raise ScriptIngestionError(message) from exc
     except TimeoutError as exc:
-        # pragma: no cover - deterministic message asserted in tests
         message = "LangGraph script execution exceeded the configured timeout"
         raise ScriptIngestionError(message) from exc
     except Exception as exc:  # pragma: no cover - exercised via tests
